Change some EXPECT_EQ to ASSERT_EQ in test_action_server.
There are several places where we run a function, check the
result code, and then look at some of the output variables.
In these tests, we should always use ASSERT_EQ for checking
the return code; if that fails, we shouldn't go on and do
the rest of the tests, as we will be testing uninitialized
data at that point.  Go through test_action_server and make
these changes as appropriate.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>