Clarify storing of current_time. (#850)
clang static analysis (incorrectly) assumes that the value
of "storage->active" can change between the first and second
branch statements.  While we know that isn't actually true,
it is more clear and the same amount of code to just have
the branch once.  Rearrange the code here so that we collapse
all of the storage->active parts into one if block.

There should be no functional change with this.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>