Skip the test on CentOS.
Instead of trying to fix the test on CentOS, just skip it.
This relies on a file called /etc/os-release being available
(which it is on CentOS and Ubuntu).  If we see that file, we
parse it and find the ID.  If we see that the ID is "centos"
(including the quotes), we skip compiling and running the
test_node tests.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>