Micro-optimizations in rcl (#965)
* Only check wait_set->impl for NULL pointer.

The previous line already checked wait_set for NULL pointer,
so there is no reason to call rcl_wait_set_is_valid and
recheck that pointer.

* Remove some debug statements from rcl_wait.

While profiling an application, I found that a lot of time was
spent just checking whether a log message should be output
or not based on the current debug level.  None of the individual
calls are expensive, but since rcl_wait is called so often,
it ends up showing up in the profile.

This patch somewhat controversially removes the debug statements
from rcl_wait().  My view on it is that the utility of these
is fairly low (if you ever turned them on, you would be flooded
with information), and the cost is relatively high.  If you
are really debugging this stuff, you can add in your own, more
targeted debug statements or use a debugger.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>