Shorten wait time for a subscription not being ready. (#1092)
We shouldn't wait around for 10 seconds for a subscription
to not be ready; that can take literally forever.  Instead,
just wait one second.  While we are in here, also make
the case where we *are* waiting for a subscription to be
ready more reactive by reducing the wait time to 100ms,
with more tries.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>