Fix a clang warning for suspicious string concatentation.
clang warns when strings are concatenated within array
declarations; this can help to catch bugs.

However, in the cases here we actually mean to do string
concatenation.  Signal this to the compiler by making a
separate variable where we do the concatenation, which
quiets the clang warning and should have no effect
on operation.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>