Improve the reliability of test_get_type_description_service. (#1107)
We were seeing occasional failures of test_get_type_description_service
on all platforms.

It turns out that the removal of a service from the graph
cache is an asynchronous operation.  In particular, all of
our current RMW implementations publish a graph update to
the graph topic, and only remove things from the graph once
that data has been delivered.  This can potentially happen
in another thread.

That means that immediately after service_fini(), a call to
get_service_names_and_types() may actually still have the
old service name in it.  Since our get_type_description
tests were relying on this to go away, this was causing it
to be flakey.

We fix this here by adding in a new helper function,
service_not_exists().  This helper is not just the inverse
of service_exists() (which returns immediately when it finds
the service in the graph cache).  Instead, service_not_exists()
waits until the service has *left* the cache before returning
(or times out).

In my testing, this fixes the flakiness locally.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>