Fix up rmw_cyclonedds timestamp testing. (#1156)
We are about to fix it so that rmw_cyclonedds has receive_timestamp
support, so we also need to enable that support here
in rcl.  We actually rewrite the logic a bit because now the
only combination that doesn't work is rmw_connextdds on Windows.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
(cherry picked from commit 6d53d24a863c3e9e4a41e9fe5f550271210d9d9d)