Fix up type_description tests.
In particular, make sure to match all service_description_init
with the appropriate finis.  While we are in here, also add in
a "service_not_exists" function that will quit much earlier,
speeding up the test.

This is essentially a backport of https://github.com/ros2/rcl/pull/1107
and parts of https://github.com/ros2/rcl/pull/1112, but adapted
for Iron.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>